task7 completed
the report entity was created, as well as its additional entities. a repository and a service for working with it were implemented. asynchronous multithreading was implemented in 1 of the report generation service methods. rest controller has been implemented to work with reports. the basic path for working with reports is '/reports'.